+++ R/run_engine.R
@@ -38,6 +38,16 @@ run_engine <- function(arm_list,
   env_setup_pt <- input_list$env_setup_pt
   env_setup_arm <- input_list$env_setup_arm
   debug <- input_list$debug
+  
+  if(input_list$accum_backwards){
+    input_list$ongoing_inputs_lu <- paste0(input_list$uc_lists$ongoing_inputs,"_lastupdate",recycle0 = TRUE)
+    input_out_v <- c(input_list$input_out,
+                     input_list$ongoing_inputs_lu
+    )
+  }else{
+    input_out_v <- c(input_list$input_out)
+  }
+  
 
   #1 Loop per patient ----------------------------------------------------------
   patdata <- vector("list", length=npats) # empty list with npats elements
@@ -80,17 +90,7 @@ run_engine <- function(arm_list,
     
     #Extract the inputs that are common for each patient across interventions
     if(!is.null(common_pt_inputs)){
-      
-      if(env_setup_pt){
-        load_inputs2(inputs = input_list_pt,list_uneval_inputs = common_pt_inputs)
-      } else{
-        input_list_pt <- as.environment(
-          load_inputs(inputs = as.list(input_list_pt),
-                      list_uneval_inputs = common_pt_inputs)
-          )
-        parent.env(input_list_pt) <- parent.env(input_list)
-        
-      }
+        load_inputs(inputs = input_list_pt,list_uneval_inputs = common_pt_inputs)
       
       
       if(input_list$debug){ 
@@ -138,16 +138,7 @@ run_engine <- function(arm_list,
                     simulation=input_list$simulation, patient_id=i, arm=arm, .warden_ctx = .warden_ctx)
       
       if(!is.null(unique_pt_inputs)){
-        
-        if(env_setup_arm){
-          load_inputs2(inputs = input_list_arm,list_uneval_inputs = unique_pt_inputs)
-        } else{
-          input_list_arm <- as.environment(
-            load_inputs(inputs = as.list(input_list_arm),
-                        list_uneval_inputs = unique_pt_inputs)
-            )
-          parent.env(input_list_arm) <- parent.env(input_list_pt)
-        }
+          load_inputs(inputs = input_list_arm,list_uneval_inputs = unique_pt_inputs)
         
         if(input_list_pt$debug){ 
           dump_info <- debug_inputs(input_list_pt,input_list_arm)
@@ -220,32 +211,14 @@ run_engine <- function(arm_list,
       this_patient[[arm]]$evtlist <- NULL
 
       input_list_arm$curtime <- 0
-      
-      if(input_list$accum_backwards){
-        input_list_arm$ongoing_inputs_lu <- paste0(input_list_arm$uc_lists$ongoing_inputs,"_lastupdate",recycle0 = TRUE)
-        input_out_v <- c(input_list_arm$input_out,
-                         input_list_arm$ongoing_inputs_lu
-        )
-      }else{
-        input_out_v <- c(input_list_arm$input_out)
-      }
 
       n_evt <- 0
-   
-      if(input_list$accum_backwards){
-        input_list_arm$ongoing_inputs_lu <- paste0(input_list_arm$uc_lists$ongoing_inputs,"_lastupdate",recycle0 = TRUE)
-        inputs_out_v <- c(input_list_arm$input_out,
-                          input_list_arm$ongoing_inputs_lu
-        )
-      }else{
-        inputs_out_v <-  input_list_arm$input_out
-      }
 
       # Update the event queue reference for new_event, modify_event, etc.
       # Note that we are only assigning the pointer, so any changes to input_list_arm$cur_evtlist
       # will equally affect event_queue
       assign("cur_evtlist", event_queue, envir = input_list_arm)
-
+      
       while (!queue_empty(event_queue)) {
         if(is.infinite(next_event(1,event_queue)$time)){
           break
@@ -271,12 +244,12 @@ run_engine <- function(arm_list,
                                          evttime = current_time),
                                     arm,
                                     input_list_arm)
-        
+       
         #Get extra objects to be exported
-        if(is.null(inputs_out_v)){
+        if(is.null(input_out_v)){
           extra_data <- list()
         } else{
-          extra_data <-  mget(inputs_out_v, input_list_arm) 
+          extra_data <-  mget(input_out_v, input_list_arm) 
         }
         extra_data <- extra_data[!vapply(extra_data, is.null, TRUE)]